refactor: split routes and services
